test(core): extract shared kdbx create_test_database fixture
SonarQube flagged a 68-line duplicated block across the entries,
favicons, and custom_icons test modules — each carried a near-identical
`create_test_database` helper (differing only in the db filename and
display name). Adding the new expiry fields to those literals tipped the
block over the new-code duplication threshold.

Extract a single `#[cfg(test)]` fixture into `kdbx::test_support` and
have the three test modules import it, removing the copies. No behavior
change; the helper still seeds "Entry A"/"Entry B" the tests rely on.
